# split php source to array to speed up lexing, closes #70