Significant decrease in parsing time, especially noticable in large files. When testing example file from #70 in Chrome 55, time decrease from 7s to 250ms and in Waterfox from 6.2s to 210ms