@Hirse 2017-01-22
Thanks for improvement. I had to change the `split` to use a regex to work on Windows.